+++ ICD.Connect.Sources.TvTuner/Streaming/IStreamTvTunerDevice.cs
@@ -1,30 +1,8 @@
-using System;
-using ICD.Common.Properties;
-using ICD.Common.Utils.EventArguments;
-using ICD.Connect.Devices;
+using ICD.Connect.Routing.Devices.Streaming;
 
 namespace ICD.Connect.Sources.TvTuner.Streaming
 {
-	public interface IStreamTvTunerDevice : IDevice
+	public interface IStreamTvTunerDevice : IStreamSourceDevice
 	{
-		/// <summary>
-		/// Raised when the stream URI changes
-		/// </summary>
-		[PublicAPI]
-		event EventHandler<GenericEventArgs<Uri>> OnStreamUriChanged;
-
-		/// <summary>
-		/// URI of the stream
-		/// </summary>
-		[PublicAPI]
-		[CanBeNull]
-		Uri StreamUri { get; }
-
-		/// <summary>
-		/// Goes to the given channel stream
-		/// </summary>
-		/// <param name="uri"></param>
-		[PublicAPI]
-		void SetStreamUri(Uri uri);
 	}
 }